% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departamento.R
\docType{data}
\name{departamento}
\alias{departamento}
\title{Población, homicidios y tasa de homicidios.}
\format{
Un objeto de la clase \code{data.frame}
}
\source{
\href{https://www.ine.gob.hn/V3/}{Instituto Nacional de Estadísticas de Honduras (INE)}

\href{https://www.sepol.hn/}{Sistema Estadistico Policial en Linea (SEPOL)}
}
\usage{
data(departamento)
}
\description{
Datos proporcioandos por el Instituto Nacional de Estadística Honduras (INE) y
el Sistema Estadistico Policial en Linea (SEPOL) que incluye la población, homicidios y
tasa de homicidios por 100,000 habitantes en los 18 departamentos de Honduras en 2019.
}
\details{
\itemize{
\item departamento: Nombre del departamento.
\item poblacion: Población.
\item homicidios: Total de homicidios.
\item tasa_homicidios: Tasa de homicidios por 100,000 habitantes.
}
}
\examples{
data(departamento)
head(departamento)

}
\keyword{datasets}
