% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_departamental.R
\docType{data}
\name{map_departamental}
\alias{map_departamental}
\title{Mapa de la división política departamental}
\format{
Un objeto de la clase \code{data.frame}
}
\source{
\href{https://datos.gob.hn/}{Gobierno de Honduras}
}
\usage{
data(map_departamental)
}
\description{
Datos proporcionados por la Iniciativa de Datos Abiertos del Gobierno de Honduras que
proporciona información de los limites administrativos de los departamento de Honduras.
}
\details{
\itemize{
\item GEOCODIGO: .
\item NOMBRE:
\item Hectares:
\item geometry: .
}
}
\examples{
data(map_departamental)
tibble::glimpse(map_departamental)

}
\keyword{datasets}
