% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_municipal.R
\docType{data}
\name{map_municipal}
\alias{map_municipal}
\title{Mapa de la división política municipal}
\format{
Un objeto de la clase \code{data.frame}
}
\source{
\href{https://datos.gob.hn/}{Gobierno de Honduras}
}
\usage{
data(map_municipal)
}
\description{
Datos proporcionados por la Iniciativa de Datos Abiertos del Gobierno de Honduras que
proporciona información de los limite administrativos municipales de Honduras.
}
\details{
\itemize{
\item GEOCODIGO: .
\item NOMBRE:
}
}
\examples{
data(map_municipal)
str(map_municipal)

}
\keyword{datasets}
