% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openhn-package.R, R/openhn.R
\docType{package}
\name{openhn-package}
\alias{openhn}
\alias{openhn-package}
\title{openhn: Datos económicos y sociales de Honduras.}
\description{
Conjunto de datos económicos y sociales de Honduras los cuales son
    obtenidos de las principales instituciones económicas del país y de organismos internacionales.
    Ordenados en formato Tidy.

Conjunto de datos económicos y sociales de Honduras los cuales son
    obtenidos de las principales instituciones económicas del país y de organismos internacionales.
    Ordenados en formato Tidy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nolivera007}
}


Useful links:
\itemize{
  \item \url{https://github.com/Nolivera007}
}

}
\keyword{internal}
