% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produccion.R
\docType{data}
\name{produccion}
\alias{produccion}
\title{Producción por actividad económica.}
\format{
Un objeto de la clase \code{data.frame}
}
\source{
\href{https://www.bch.hn/}{Banco Central de Honduras (BCH)}
}
\usage{
data(produccion)
}
\description{
Datos proporcionados por el Banco Central de Honduras (BCH) que incluye
la producción (millones de lempiras)
de las actividades económicas trimestral en valores constantes.
}
\details{
\itemize{
\item id: número del 1 al 15.
\item actividad_economica: Concepto.
\item trimestre: Trimestre.
\item hnl: Millones de Lempiras.
}
}
\examples{
data(produccion)
head(produccion)

}
\keyword{datasets}
