% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departamento.R
\docType{data}
\name{departamento}
\alias{departamento}
\title{Población, homicidios y tasa de homicidios.}
\format{
Un objeto de la clase \code{"data.frame"}
}
\source{
Instituto Nacional de Estadísticas de Honduras (INE)
\url{}
}
\usage{
data(departamento)
}
\description{
\itemize{
\item departamento: 18 departamentos de Honduras.
\item poblacion: Población.
\item homicidios: Total de homcidios.
\item tasa_homicidios: Tasa de homicidios por 100,000 habitantes.
}
}
\examples{
Cálcule la media, la media recortada y la mediana de la población

library(dplyr)
library(openhn)

data(departamento)

departamento \%>\%
summarise(
 media = mean(poblacion),
 media_recortada = mean(poblacion, trim = 0.1),
 mediana = median(poblacion)
)

}
\keyword{datasets}
