% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departamento.R
\docType{data}
\name{departamento}
\alias{departamento}
\title{Población, homicidios y tasa de homicidios.}
\format{
Un objeto de la clase \code{"data.frame"}
}
\source{
Instituto Nacional de Estadísticas de Honduras (INE)
\url{https://www.ine.gob.hn/V3/}

Sistema Estadistico Policial en Linea (SEPOL)
\url{https://www.sepol.hn/}
}
\usage{
data(departamento)
}
\description{
Población, homicidios y tasa de homicidios por 100,000
habitantes en los 18 departamentos de Honduras en 2019.
Datos del Instituto Nacional de Estadísticas de Honduras y
el Sistema Estadistico Policial en Linea.
}
\details{
\itemize{
\item departamento: 18 departamentos de Honduras.
\item poblacion: Población.
\item homicidios: Total de homcidios.
\item tasa_homicidios: Tasa de homicidios por 100,000 habitantes.
}
}
\examples{
1. Cálcule la media y la mediana de la población.

library(dplyr)
library(openhn)

data(departamento)
head(departamento)

departamento \%>\%
summarise(
 media = mean(poblacion),
 mediana = median(poblacion)
)

}
\keyword{datasets}
