% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openhn-package.R
\docType{package}
\name{openhn-package}
\alias{openhn}
\alias{openhn-package}
\title{openhn: Datos económicos y sociales de Honduras de fuentes públicas.}
\description{
El objetivo del paquete openhn es proporcionar a investigadores, analistas y cualquier persona interesa en datos socioeconómicos de Honduras a nivel departamental y municipal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nolivera007/openhn}
}

}
\keyword{internal}
