% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openhn-package.R, R/openhn.R
\docType{package}
\name{openhn}
\alias{openhn}
\title{Datos económicos y sociales de Honduras.}
\description{
Este paquete contiene una serie de datos económicos y sociales
de Honduras.

Este paquete contiene una serie de datos económicos y sociales
de Honduras.
}
\seealso{
\code{\link{pib_produccion}}
}
